百度审核接口功能编写

+++ src/main/java/com/pxene/pap/repository/basic/view/CreativeImageDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CreativeImageDao {
-    int countByExample(CreativeImageModelExample example);
+    long countByExample(CreativeImageModelExample example);
 
     int deleteByExample(CreativeImageModelExample example);
