+++ src/main/java/com/pxene/pap/repository/basic/view/CreativeVideoDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface CreativeVideoDao {
-    int countByExample(CreativeVideoModelExample example);
+    long countByExample(CreativeVideoModelExample example);
 
     int deleteByExample(CreativeVideoModelExample example);
